feat: logic to move to modify ticket on assigned ticket
